Restructure to allow future submodules create one-line python interface
